peripheral-driver/requirements.txt
evdev==1.6.1
RPi.GPIO==0.7.1
pygame==2.5.2 